#112 活動確認フォームのレイアウト
