+++ frontend/src/App.vue
@@ -23,15 +23,8 @@
           <li class="nav-item">
             <router-link class="nav-link" to="/register/todo">Todo登録</router-link>
           </li>
-          <li class="nav-item dropdown">
-            <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
-              実績確認
-            </a>
-            <ul class="dropdown-menu">
-              <li class="dropdown-item"><router-link class="nav-link" style="color: black;" to="/view/month-activities">月別</router-link></li>
-              <li class="dropdown-item"><router-link class="nav-link" style="color: black;" to="/view/year-activities">年別</router-link></li>
-              <li class="dropdown-item"><router-link class="nav-link" style="color: black;" to="/view/all-activities">全期間</router-link></li>
-            </ul>
+          <li class="nav-item">
+            <router-link class="nav-link" to="/view/activity/">実績確認</router-link>
           </li>
           <li class="nav-item">
             <router-link class="nav-link" to="/user/info">ユーザー情報</router-link>
@@ -63,14 +56,8 @@
         <li class="nav-item" data-bs-dismiss="offcanvas">
           <router-link class="nav-link" to="/register/todo">Todo登録</router-link>
         </li>
-        <li class="nav-item dropdown">
-          <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
-            実績確認
-          </a>
-          <ul class="dropdown-menu" data-bs-dismiss="offcanvas">
-            <li class="dropdown-item"><router-link class="nav-link" style="color: black;" to="/view/month-activities">月別</router-link></li>
-            <li class="dropdown-item"><router-link class="nav-link" style="color: black;" to="/view/all-activities">全期間</router-link></li>
-          </ul>
+        <li class="nav-item" data-bs-dismiss="offcanvas">
+          <router-link class="nav-link" to="/view/activity/">実績確認</router-link>
         </li>
         <li class="nav-item" data-bs-dismiss="offcanvas">
           <router-link class="nav-link" to="/user/info">ユーザー情報</router-link>
@@ -85,6 +72,7 @@
     </div>
   </nav>
   <p v-if="logoutMsg" class="logoutMsg">{{ logoutMsg }}</p>
+  <br>
   <BContainer>
     <router-view></router-view>
   </BContainer>
