+++ frontend/src/router/index.js
@@ -3,14 +3,12 @@ import RegisterUser from '../views/RegisterUser.vue'
 import LoginView from '../views/LoginForm.vue'
 import UserHomeView from '../views/UserHome.vue'
 import RegisterSalary from '../views/RegisterSalary.vue'
-import MonthlyInfo from '../views/MonthlyInfo.vue'
 import RegisterTodo from '../views/RegisterTodo.vue'
 import InquiryForm from '../views/InquiryForm.vue'
-import AllPeriodInfo from '../views/AllPeriodInfo.vue'
 import EditTodo from '../views/EditTodo.vue'
 import UserInfo from '../views/UserInfo.vue'
-import YearInfo from '../views/YearInfo.vue'
 import ActivityHome from '../views/ActivityHome.vue'
+import ActivityInfo from '../views/ActivityInfo.vue'
 
 const routes = [
   {
@@ -34,11 +32,6 @@ const routes = [
     component: RegisterSalary
   },
   {
-    path:"/view/month-activities",
-    name:"MonthlyInfo",
-    component:MonthlyInfo
-  },
-  {
     path:"/register/todo",
     name:"RegisterTodo",
     component:RegisterTodo
@@ -47,11 +40,6 @@ const routes = [
     path:"/register/inquiry",
     name:"InquiryForm",
     component:InquiryForm
-  },
-  {
-    path:"/view/all-activities",
-    name:"AllPeriodInfo",
-    component:AllPeriodInfo
   },
   {
     path:"/edit/todo",
@@ -64,14 +52,14 @@ const routes = [
     component:UserInfo
   },
   {
-    path:"/view/year-activities",
-    name:"YearInfo",
-    component:YearInfo
-  },
-  {
     path:"/register/activity",
     name:"ActivityHomw",
     component:ActivityHome
+  },
+  {
+    path:"/view/activity",
+    name:"ActivityInfo",
+    component:ActivityInfo
   }
 ]
