+++ frontend/src/views/lib/error.js
@@ -120,3 +120,34 @@ export function getMonthlyinfoError(response, activities, message, router){
         handleError
     }
 }
+
+export function allActivitiesError(message, router){
+    const handleError = async(error) =>{
+        if (error.response){
+            switch (error.response.status){
+                case 401:
+                    router.push(
+                        {"path":"/login",
+                        "query":{message:"再度ログインしてください"}
+                    })
+                    break;
+                case 422:
+                    message.value = error.response.data.detail;
+                    break;
+                case 500:
+                    message.value =  "情報の取得に失敗しました"
+                    break;
+                default:
+                    message.value = error.response.data.detail;
+                }
+            } else if (error.request){
+                message.value =  "リクエストがサーバーに到達できませんでした"
+            } else {
+                message.value =  "不明なエラーが発生しました。管理者にお問い合わせください"
+        }
+    }
+
+    return {
+        handleError
+    }
+}
